+++ lib/pigeon/adm/worker.ex
@@ -1,7 +1,7 @@
-defmodule Pigeon.ADMWorker do
+defmodule Pigeon.ADM.Worker do
   @moduledoc """
-    Handles all Amazon ADM request and response parsing.
-    Includes managing OAuth2 tokens.
+  Handles all Amazon ADM request and response parsing.
+  Includes managing OAuth2 tokens.
   """
   use GenServer
   require Logger
