+++ test/fcm/worker_test.exs
@@ -1,16 +1,16 @@
-defmodule Pigeon.GCMWorkerTest do
+defmodule Pigeon.FCM.WorkerTest do
   use ExUnit.Case
-  alias Pigeon.GCMWorker
-  alias Pigeon.GCM.NotificationResponse
+  alias Pigeon.FCM
+  alias Pigeon.FCM.{NotificationResponse}
 
   @data %{"message" => "Test push"}
   @payload %{"data" => @data}
 
-  defp valid_gcm_reg_id, do: Application.get_env(:pigeon, :test)[:valid_gcm_reg_id]
+  defp valid_fcm_reg_id, do: Application.get_env(:pigeon, :test)[:valid_fcm_reg_id]
 
   test "parse_result with success" do
     {:ok, response} =
-      GCMWorker.parse_result1(
+      FCM.Worker.parse_result1(
         ["regid"],
         [%{ "message_id" => "1:0408" }],
         &(&1), %NotificationResponse{}
@@ -20,7 +20,7 @@ defmodule Pigeon.GCMWorkerTest do
 
   test "parse_result with success and new registration_id" do
     {:ok, response} =
-      GCMWorker.parse_result1(
+      FCM.Worker.parse_result1(
         ["regid"],
         [%{ "message_id" => "1:2342", "registration_id" => "32" }],
         &(&1), %NotificationResponse{}
@@ -32,7 +32,7 @@ defmodule Pigeon.GCMWorkerTest do
 
   test "parse_result with error unavailable" do
     {:ok, response} =
-      GCMWorker.parse_result1(
+      FCM.Worker.parse_result1(
         ["regid"],
         [%{ "error" => "Unavailable" }],
         &(&1),
@@ -43,7 +43,7 @@ defmodule Pigeon.GCMWorkerTest do
 
   test "parse_result with custom error" do
     {:ok, response} =
-      GCMWorker.parse_result1(
+      FCM.Worker.parse_result1(
         ["regid"],
         [%{ "error" => "CustomError" }],
         &(&1),
@@ -53,11 +53,13 @@ defmodule Pigeon.GCMWorkerTest do
   end
 
   test "send malformed JSON" do
-    {:ok, pid} = GCMWorker.start_link(:gonecrashing, key: Application.get_env(:pigeon, :gcm)[:key])
+    #{:ok, pid} = FCM.Worker.start_link(:gonecrashing, key: Application.get_env(:pigeon, :fcm)[:key])
+    opts = [name: :gonecrashing, key: Application.get_env(:pigeon, :test)[:fcm_key]]
+    {:ok, pid} = FCM.start_connection(opts)
+
     me = self()
-    :gen_server.cast(pid, {:push, :gcm, {"toto", "this is not json"}, &(send me, &1), %{}})
+    :gen_server.cast(pid, {:push, :fcm, {"toto", "this is not json"}, &(send me, &1), %{}})
     assert_receive {:error, :malformed_json}, 5000
     :gen_server.cast(pid, :stop)
   end
-
 end
