+++ test/gcm_test.exs
@@ -1,23 +1,22 @@
-defmodule Pigeon.GCMTest do
+defmodule Pigeon.FCMTest do
   use ExUnit.Case
-  doctest Pigeon.GCM.Notification
+  doctest Pigeon.FCM.Notification
 
-  alias Pigeon.GCM.Notification
-  alias Pigeon.GCM.NotificationResponse
+  alias Pigeon.FCM.{Notification, NotificationResponse}
   require Logger
 
   @data %{"message" => "Test push"}
   @payload %{"data" => @data}
 
-  defp valid_gcm_reg_id, do: Application.get_env(:pigeon, :test)[:valid_gcm_reg_id]
+  defp valid_fcm_reg_id, do: Application.get_env(:pigeon, :test)[:valid_fcm_reg_id]
 
   test "successfully sends a valid push" do
     {:ok, notification} =
-      valid_gcm_reg_id()
+      valid_fcm_reg_id()
       |> Notification.new(%{}, @data)
-      |> Pigeon.GCM.push
+      |> Pigeon.FCM.push
 
-    assert notification.ok == [valid_gcm_reg_id()]
+    assert notification.ok == [valid_fcm_reg_id()]
   end
 
   test "Can merge two responses" do
@@ -31,7 +30,7 @@ defmodule Pigeon.GCMTest do
       update: ["12"],
       error: %{"error" => ["2"], "error2"=> ["1"]}
     }
-    assert Pigeon.GCM.merge(nr1, nr2) ==
+    assert Pigeon.FCM.merge(nr1, nr2) ==
       %NotificationResponse{
         ok: ["42", "43"],
         retry: ["12"],
@@ -43,13 +42,13 @@ defmodule Pigeon.GCMTest do
   test "Message for less than 1000 recipients should not be chunked" do
     regs = Enum.to_list(1..999)
     notification = Notification.new(regs, %{}, @data)
-    assert [{^regs, encoded}] = res = Pigeon.GCM.encode_requests(notification)
+    assert [{^regs, encoded}] = res = Pigeon.FCM.encode_requests(notification)
   end
 
   test "Message for over 1000 recipients should be chunked" do
     regs = Enum.to_list(1..2534)
     notification = Notification.new(regs, %{}, @data)
-    res = Pigeon.GCM.encode_requests(notification)
+    res = Pigeon.FCM.encode_requests(notification)
     assert [{r1, e1}, {r2, e2}, {r3, e3}]  = res
     assert length(r1) == 1000
     assert length(r2) == 1000
@@ -58,18 +57,18 @@ defmodule Pigeon.GCMTest do
 
   test "successfully sends a valid push with an explicit config" do
     response =
-      valid_gcm_reg_id()
+      valid_fcm_reg_id()
       |> Notification.new(%{}, @data)
-      |> Pigeon.GCM.push(%{gcm_key: "explicit"})
+      |> Pigeon.FCM.push(%{key: "explicit"})
 
      assert response == {:error, :unauthorized}
   end
 
   test "successfully sends a valid push with callback" do
-    reg_id = valid_gcm_reg_id()
+    reg_id = valid_fcm_reg_id()
     n = Notification.new(reg_id, %{}, @data)
     pid = self()
-    Pigeon.GCM.send_push(n, fn(x) -> send pid, x end, %{})
+    Pigeon.FCM.send_push(n, fn(x) -> send pid, x end, %{})
 
     assert_receive {:ok, notification}, 5000
     assert notification.ok == [reg_id]
@@ -79,9 +78,9 @@ defmodule Pigeon.GCMTest do
     reg_id = "bad_registration_id"
     n = Notification.new(reg_id, %{}, @data)
     pid = self()
-    Pigeon.GCM.send_push(n, fn(x) -> send pid, x end, %{})
+    Pigeon.FCM.send_push(n, fn(x) -> send pid, x end, %{})
 
-    assert_receive {:ok, %Pigeon.GCM.NotificationResponse{remove: ["bad_registration_id"]}}, 5000
+    assert_receive {:ok, %Pigeon.FCM.NotificationResponse{remove: ["bad_registration_id"]}}, 5000
     assert n.registration_id == reg_id
     assert n.payload == %{"data" => @data}
   end
@@ -89,7 +88,7 @@ defmodule Pigeon.GCMTest do
   test "encode_requests with one registration_id" do
     registration_id = "123456"
     payload = Notification.new(registration_id, %{},@data)
-    assert Pigeon.GCM.encode_requests(payload) ==
+    assert Pigeon.FCM.encode_requests(payload) ==
       [{["123456"], ~S({"to":"123456","priority":"normal","data":{"message":"Test push"}})}]
   end
 
@@ -97,7 +96,7 @@ defmodule Pigeon.GCMTest do
     registration_id = ["aaaaaa", "bbbbbb", "cccccc"]
     payload = Notification.new(registration_id, %{},@data)
     expected = ~S({"registration_ids":["aaaaaa","bbbbbb","cccccc"],"priority":"normal","data":{"message":"Test push"}})
-    assert Pigeon.GCM.encode_requests(payload) == [{registration_id, expected}]
+    assert Pigeon.FCM.encode_requests(payload) == [{registration_id, expected}]
   end
 
 end
